backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-multipart
python-dotenv
requests
beautifulsoup4
playwright
spacy
sentence-transformers
scikit-learn
redis
passlib[bcrypt]
python-jose[cryptography]
aiohttp
pymupdf
python-docx
